[fuchsia] Remove Dart CF v1 runner.
This runner is no longer used anywhere, it has been fully
replaced with the CF v2 runner.

The Flutter CF v1 runner is still in-use.
