+++ ci/licenses_golden/licenses_flutter
@@ -2111,12 +2111,10 @@ FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/builtin_libraries.cc
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/builtin_libraries.h
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_component_controller.cc
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_component_controller.h
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_component_controller_v2.cc
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_component_controller_v2.h
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_runner.cc
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_runner.h
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_test_component_controller_v2.cc
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_test_component_controller_v2.h
+FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_test_component_controller.cc
+FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/dart_test_component_controller.h
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/embedder/builtin.dart
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/embedder/script_runner_snapshot.dart
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/embedder/shim.dart
@@ -2126,19 +2124,11 @@ FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/embedder/snapshot.h
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/kernel/libraries.json
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/logging.h
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/main.cc
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/aot_product_runtime
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/aot_runtime
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/common.shard.cml
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_aot_product_runner.cml
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_aot_product_runner.cmx
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_aot_runner.cml
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_aot_runner.cmx
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_jit_product_runner.cml
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_jit_product_runner.cmx
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_jit_runner.cml
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/dart_jit_runner.cmx
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/jit_product_runtime
-FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/meta/jit_runtime
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/service_isolate.cc
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/service_isolate.h
 FILE: ../../../flutter/shell/platform/fuchsia/dart_runner/vmservice/empty.dart
