+++ shell/platform/fuchsia/dart_runner/README.md
@@ -1,3 +1,4 @@
 # Dart Application Runner
 
-An Runner that runs Dart programs.
+Implements the `fuchsia::component::runner::ComponentRunner` FIDL interface to
+launch and run Dart applications that don't use Flutter.
