+++ shell/platform/fuchsia/dart_runner/dart_runner.h
@@ -10,24 +10,17 @@
 #include <lib/fidl/cpp/binding_set.h>
 #include <lib/sys/cpp/component_context.h>
 
-#include "dart_test_component_controller_v2.h"
+#include "dart_test_component_controller.h"
 #include "runtime/dart/utils/mapped_resource.h"
 
 namespace dart_runner {
 
-class DartRunner : public fuchsia::sys::Runner,
-                   public fuchsia::component::runner::ComponentRunner {
+class DartRunner : public fuchsia::component::runner::ComponentRunner {
  public:
   explicit DartRunner(sys::ComponentContext* context);
   ~DartRunner() override;
 
  private:
-  // |fuchsia::sys::Runner| implementation:
-  void StartComponent(fuchsia::sys::Package package,
-                      fuchsia::sys::StartupInfo startup_info,
-                      fidl::InterfaceRequest<fuchsia::sys::ComponentController>
-                          controller) override;
-
   // |fuchsia::component::runner::ComponentRunner| implementation:
   void Start(
       fuchsia::component::runner::ComponentStartInfo start_info,
@@ -36,13 +29,12 @@ class DartRunner : public fuchsia::sys::Runner,
 
   // Add test components to this map to ensure it is kept alive in memory for
   // the duration of test execution and retrieval of exit code.
-  std::map<DartTestComponentControllerV2*,
-           std::shared_ptr<DartTestComponentControllerV2>>
+  std::map<DartTestComponentController*,
+           std::shared_ptr<DartTestComponentController>>
       test_components_;
 
   // Not owned by DartRunner.
   sys::ComponentContext* context_;
-  fidl::BindingSet<fuchsia::sys::Runner> bindings_;
   fidl::BindingSet<fuchsia::component::runner::ComponentRunner>
       component_runner_bindings_;
