--- shell/platform/fuchsia/dart_runner/meta/aot_runtime
@@ -1,3 +1,0 @@
-{
-    "runner": "fuchsia-pkg://fuchsia.com/dart_aot_runner#meta/dart_aot_runner.cmx"
-}
