--- shell/platform/fuchsia/dart_runner/meta/dart_aot_product_runner.cmx
@@ -1,20 +1,0 @@
-{
-    "program": {
-        "data": "data"
-    },
-    "sandbox": {
-        "features": [
-            "config-data",
-            "root-ssl-certificates"
-        ],
-        "services": [
-            "fuchsia.device.NameProvider",
-            "fuchsia.feedback.CrashReporter",
-            "fuchsia.intl.PropertyProvider",
-            "fuchsia.logger.LogSink",
-            "fuchsia.net.name.Lookup",
-            "fuchsia.posix.socket.Provider",
-            "fuchsia.tracing.provider.Registry"
-        ]
-    }
-}
