--- shell/platform/fuchsia/dart_runner/meta/dart_jit_runner.cmx
@@ -1,23 +1,0 @@
-{
-    "program": {
-        "data": "data"
-    },
-    "sandbox": {
-        "features": [
-            "config-data",
-            "deprecated-ambient-replace-as-executable",
-            "root-ssl-certificates"
-        ],
-        "services": [
-            "fuchsia.device.NameProvider",
-            "fuchsia.feedback.CrashReporter",
-            "fuchsia.intl.PropertyProvider",
-            "fuchsia.logger.LogSink",
-            "fuchsia.net.name.Lookup",
-            "fuchsia.posix.socket.Provider",
-            "fuchsia.process.Launcher",
-            "fuchsia.process.Resolver",
-            "fuchsia.tracing.provider.Registry"
-        ]
-    }
-}
