--- shell/platform/fuchsia/dart_runner/meta/jit_runtime
@@ -1,3 +1,0 @@
-{
-    "runner": "fuchsia-pkg://fuchsia.com/dart_jit_runner#meta/dart_jit_runner.cmx"
-}
