+++ shell/platform/fuchsia/flutter/README.md
@@ -1,4 +1,5 @@
 Flutter Application Runner
 ==========================
 
-Implements the `fuchsia::sys::Runner` FIDL interface to launch and run multiple Flutter applications within the same process.
+Implements the `fuchsia::component::runner::ComponentRunner` FIDL interface to
+launch and run multiple Flutter applications within the same process in Fuchsia.
