+++ tools/fuchsia/dart/dart_build_config.gni
@@ -7,7 +7,6 @@ import("//flutter/common/config.gni")
 # Builds the component in a non-product JIT build. This will
 # launch the vm service in the runner.
 dart_debug_build_cfg = {
-  runtime_meta = "//flutter/shell/platform/fuchsia/dart_runner/meta/jit_runtime"
   runner_dep = "//flutter/shell/platform/fuchsia/dart_runner:dart_jit_runner"
   platform_name = "dart_runner"
   is_aot = false
@@ -20,7 +19,6 @@ dart_debug_build_cfg = {
 # This configuration is not compatible with a --release build since the
 # profile aot runner is built without asserts.
 dart_aot_debug_build_cfg = {
-  runtime_meta = "//flutter/shell/platform/fuchsia/dart_runner/meta/aot_runtime"
   runner_dep = "//flutter/shell/platform/fuchsia/dart_runner:dart_aot_runner"
   platform_name = "dart_runner"
   is_aot = true
@@ -31,7 +29,6 @@ dart_aot_debug_build_cfg = {
 # Builds the component in a non-product AOT build. This will
 # launch the vm service in the runner.
 dart_profile_build_cfg = {
-  runtime_meta = "//flutter/shell/platform/fuchsia/dart_runner/meta/aot_runtime"
   runner_dep = "//flutter/shell/platform/fuchsia/dart_runner:dart_aot_runner"
   platform_name = "dart_runner"
   is_aot = true
@@ -42,8 +39,6 @@ dart_profile_build_cfg = {
 # Builds the component in a product JIT build. This will
 # not launch the vm service in the runner.
 dart_jit_release_build_cfg = {
-  runtime_meta =
-      "//flutter/shell/platform/fuchsia/dart_runner/meta/jit_product_runtime"
   runner_dep =
       "//flutter/shell/platform/fuchsia/dart_runner:dart_jit_product_runner"
   platform_name = "dart_runner"
@@ -55,8 +50,6 @@ dart_jit_release_build_cfg = {
 # Builds the component in a product AOT build. This will
 # not launch the vm service in the runner.
 dart_release_build_cfg = {
-  runtime_meta =
-      "//flutter/shell/platform/fuchsia/dart_runner/meta/aot_product_runtime"
   runner_dep =
       "//flutter/shell/platform/fuchsia/dart_runner:dart_aot_product_runner"
   platform_name = "dart_runner"
